pyproject.toml
[project]
name = "project4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
]
